detect_os back to salt_master and artifact tests
